Create case4_plants_watering.py
The system should have the following features.

The system should receive values from the soil moisture sensor. The readings are published on the MQTT server in the topic /class/stand<id>/humidity, where <id> is the number of the case with laboratory equipment.

Watering is performed at a low humidity level (sensor readings less than 15) and is accompanied by sending the message "watering" to the topic /class/stand<id>/pump, where <id> is the number of the case with laboratory equipment. It should be possible to start the pump forcibly (perform watering), upon receipt of the message "do" in the corresponding topic.

The power supply to the submersible pump is carried out using a mini-relay module or a power key.